Fix calculation due to wrong type.

+++ Extensions/MangosExtensions.cs
@@ -357,7 +357,7 @@ public static class MangosExtensions
 
     public static float GetDps(this MangosItemTemplate template)
     {
-        float divisor = template.Delay / 1000;
+        float divisor = (float)template.Delay / 1000f;
         float avgDmg = (template.DmgMin1 + template.DmgMin2 + template.DmgMin3 + template.DmgMin4 + template.DmgMin5 +
                 template.DmgMax1 + template.DmgMax2 + template.DmgMax3 + template.DmgMax4 + template.DmgMax5) / 2;
